Refactor MCP workflow planning and E2E helper logic
- Replace sharding workflow planning test delegates with kernel-focused services
- Extract workflow apply response and guidance resource hint builders
- Split MCP E2E conversation instruction, turn planning, and trace metric logic
- Add focused tests for extracted MCP workflow and E2E helpers
