+++ test/e2e/mcp/src/test/java/org/apache/shardingsphere/test/e2e/mcp/llm/conversation/LLMMCPConversationRunner.java
@@ -38,7 +38,6 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 /**
  * LLM MCP conversation runner.
@@ -89,12 +88,14 @@ public final class LLMMCPConversationRunner {
         try {
             llmChatClient.waitUntilReady();
             openInteractionClient();
+            LLMMCPConversationInstructionFactory instructionFactory = new LLMMCPConversationInstructionFactory();
+            LLMMCPConversationTurnPlanner turnPlanner = new LLMMCPConversationTurnPlanner(instructionFactory);
             boolean finalAnswerRequested = false;
             for (int turnIndex = 0; turnIndex < maxTurns; turnIndex++) {
-                finalAnswerRequested = requestFinalAnswerIfReady(scenario, messages, artifacts, finalAnswerRequested);
-                List<String> turnToolNames = finalAnswerRequested ? List.of() : createTurnToolNames(scenario, artifacts.getInteractionTrace());
-                LLMChatCompletion completion = completeTurn(scenario, messages, artifacts, finalAnswerRequested, turnToolNames);
-                Optional<LLME2EArtifactBundle> toolCallFailure = processToolCallCompletion(scenario, completion, messages, artifacts);
+                finalAnswerRequested = requestFinalAnswerIfReady(scenario, messages, artifacts, finalAnswerRequested, instructionFactory);
+                List<String> turnToolNames = finalAnswerRequested ? List.of() : turnPlanner.createTurnToolNames(scenario, artifacts.getInteractionTrace());
+                LLMChatCompletion completion = completeTurn(scenario, messages, artifacts, finalAnswerRequested, turnToolNames, instructionFactory, turnPlanner);
+                Optional<LLME2EArtifactBundle> toolCallFailure = processToolCallCompletion(scenario, completion, messages, artifacts, instructionFactory, turnPlanner);
                 if (toolCallFailure.isPresent()) {
                     return toolCallFailure.get();
                 }
@@ -103,7 +104,7 @@ public final class LLMMCPConversationRunner {
                 }
                 if (!finalAnswerRequested) {
                     if (!LLMMCPInteractionCoverage.hasRequiredInteractionCoverage(scenario.getRequiredToolNames(), artifacts.getInteractionTrace())) {
-                        messages.add(LLMChatMessage.user(createRequiredToolCallInstruction(scenario, artifacts)));
+                        messages.add(LLMChatMessage.user(instructionFactory.createRequiredToolCallInstruction(scenario, artifacts)));
                     }
                     continue;
                 }
@@ -134,18 +135,18 @@ public final class LLMMCPConversationRunner {
     }
     
     private boolean requestFinalAnswerIfReady(final LLME2EScenario scenario, final List<LLMChatMessage> messages, final LLMMCPConversationArtifacts artifacts,
-                                              final boolean finalAnswerRequested) {
+                                              final boolean finalAnswerRequested, final LLMMCPConversationInstructionFactory instructionFactory) {
         if (finalAnswerRequested || !LLMMCPInteractionCoverage.hasRequiredInteractionCoverage(scenario.getRequiredToolNames(), artifacts.getInteractionTrace())) {
             return finalAnswerRequested;
         }
-        if (hasPendingImmediateNextAction(artifacts.getInteractionTrace())) {
+        if (instructionFactory.hasPendingImmediateNextAction(artifacts.getInteractionTrace())) {
             return false;
         }
         if (scenario.getRequiredToolNames().contains("database_gateway_execute_query") && !hasExpectedExecuteQuery(scenario.getExpectedAnswer(), artifacts.getInteractionTrace())) {
-            messages.add(LLMChatMessage.user(createExpectedQueryInstruction(scenario.getExpectedAnswer())));
+            messages.add(LLMChatMessage.user(instructionFactory.createExpectedQueryInstruction(scenario.getExpectedAnswer())));
             return false;
         }
-        messages.add(LLMChatMessage.user(createFinalAnswerInstruction(scenario, artifacts.getInteractionTrace())));
+        messages.add(LLMChatMessage.user(instructionFactory.createFinalAnswerInstruction(scenario, artifacts.getInteractionTrace())));
         return true;
     }
     
@@ -169,80 +170,29 @@ public final class LLMMCPConversationRunner {
         return expectedAnswer.getSchema().equals(schema) || schema.isEmpty();
     }
     
-    private String createExpectedQueryInstruction(final LLMStructuredAnswer expectedAnswer) {
-        return String.format(Locale.ENGLISH,
-                "Required MCP tool coverage is present, but the latest successful database_gateway_execute_query did not use database `%s`, schema `%s`, and query `%s`. "
-                        + "Call database_gateway_execute_query now with exactly those arguments before returning the final JSON.",
-                expectedAnswer.getDatabase(), expectedAnswer.getSchema(), expectedAnswer.getQuery());
-    }
-    
     private LLMChatCompletion completeTurn(final LLME2EScenario scenario, final List<LLMChatMessage> messages, final LLMMCPConversationArtifacts artifacts,
-                                           final boolean finalAnswerRequested, final List<String> turnToolNames) throws IOException, InterruptedException {
-        String toolChoice = createToolChoice(scenario, artifacts, finalAnswerRequested);
-        LLMChatCompletion result = llmChatClient.complete(finalAnswerRequested ? createFinalAnswerMessages(scenario, artifacts.getInteractionTrace()) : messages,
+                                           final boolean finalAnswerRequested, final List<String> turnToolNames, final LLMMCPConversationInstructionFactory instructionFactory,
+                                           final LLMMCPConversationTurnPlanner turnPlanner) throws IOException, InterruptedException {
+        String toolChoice = turnPlanner.createToolChoice(scenario, artifacts.getInteractionTrace(), finalAnswerRequested);
+        LLMChatCompletion result = llmChatClient.complete(finalAnswerRequested ? instructionFactory.createFinalAnswerMessages(scenario, artifacts.getInteractionTrace()) : messages,
                 finalAnswerRequested ? List.of() : toolDefinitionFactory.create(turnToolNames),
                 toolChoice, finalAnswerRequested);
         artifacts.addRawModelOutput(result.getRawResponse());
         return result;
     }
     
-    private List<String> createTurnToolNames(final LLME2EScenario scenario, final List<MCPInteractionTraceRecord> interactionTrace) {
-        if (!interactionTrace.isEmpty()) {
-            String immediateActionName = findImmediateNextActionName(interactionTrace.getLast());
-            if (!immediateActionName.isEmpty() && scenario.getAllowedToolNames().contains(immediateActionName)) {
-                return List.of(immediateActionName);
-            }
-        }
-        if (hasSideEffectExecutionNextAction(interactionTrace)) {
-            List<String> readOnlyToolNames = findMissingReadOnlyToolNames(scenario, interactionTrace);
-            if (!readOnlyToolNames.isEmpty()) {
-                return List.of(readOnlyToolNames.getFirst());
-            }
-        }
-        List<String> missingToolNames = findMissingAllowedToolNames(scenario, interactionTrace);
-        return missingToolNames.isEmpty() ? scenario.getAllowedToolNames() : List.of(missingToolNames.getFirst());
-    }
-    
-    private List<String> findMissingAllowedToolNames(final LLME2EScenario scenario, final List<MCPInteractionTraceRecord> interactionTrace) {
-        return LLMMCPInteractionCoverage.findMissingRequiredInteractionNames(
-                scenario.getRequiredToolNames(), interactionTrace).stream().filter(each -> scenario.getAllowedToolNames().contains(each)).collect(Collectors.toList());
-    }
-    
-    private List<String> findMissingReadOnlyToolNames(final LLME2EScenario scenario, final List<MCPInteractionTraceRecord> interactionTrace) {
-        List<String> result = new LinkedList<>();
-        for (String each : LLMMCPInteractionCoverage.findMissingRequiredInteractionNames(scenario.getRequiredToolNames(), interactionTrace)) {
-            if (scenario.getAllowedToolNames().contains(each) && isReadOnlyToolName(each)) {
-                result.add(each);
-            }
-        }
-        return result;
-    }
-    
-    private boolean isReadOnlyToolName(final String toolName) {
-        return MCPInteractionActionNames.LIST_RESOURCES.equals(toolName)
-                || MCPInteractionActionNames.READ_RESOURCE.equals(toolName)
-                || MCPInteractionActionNames.LIST_PROMPTS.equals(toolName)
-                || MCPInteractionActionNames.GET_PROMPT.equals(toolName)
-                || MCPInteractionActionNames.COMPLETE.equals(toolName)
-                || "database_gateway_search_metadata".equals(toolName)
-                || "database_gateway_execute_query".equals(toolName);
-    }
-    
-    private String createToolChoice(final LLME2EScenario scenario, final LLMMCPConversationArtifacts artifacts, final boolean finalAnswerRequested) {
-        if (finalAnswerRequested) {
-            return "none";
-        }
-        return LLMMCPInteractionCoverage.hasRequiredInteractionCoverage(scenario.getRequiredToolNames(), artifacts.getInteractionTrace()) ? "auto" : "required";
-    }
-    
     private Optional<LLME2EArtifactBundle> processToolCallCompletion(final LLME2EScenario scenario, final LLMChatCompletion completion,
-                                                                     final List<LLMChatMessage> messages, final LLMMCPConversationArtifacts artifacts) throws InterruptedException {
+                                                                     final List<LLMChatMessage> messages, final LLMMCPConversationArtifacts artifacts,
+                                                                     final LLMMCPConversationInstructionFactory instructionFactory,
+                                                                     final LLMMCPConversationTurnPlanner turnPlanner) throws InterruptedException {
         if (completion.getToolCalls().isEmpty()) {
             return Optional.empty();
         }
         messages.add(LLMChatMessage.assistant(completion.getContent(), completion.getToolCalls()));
         for (LLMToolCall each : completion.getToolCalls()) {
-            List<String> availableToolNames = hasSideEffectExecutionNextAction(artifacts.getInteractionTrace()) ? createTurnToolNames(scenario, artifacts.getInteractionTrace()) : List.of();
+            List<String> availableToolNames = instructionFactory.hasSideEffectExecutionNextAction(artifacts.getInteractionTrace())
+                    ? turnPlanner.createTurnToolNames(scenario, artifacts.getInteractionTrace())
+                    : List.of();
             if (!availableToolNames.isEmpty() && !availableToolNames.contains(each.getName())) {
                 Map<String, Object> toolResponse = new LinkedHashMap<>(4, 1F);
                 toolResponse.put("response_mode", "tool_call_rejected");
@@ -268,7 +218,10 @@ public final class LLMMCPConversationRunner {
                 return result;
             }
         }
-        addTraceDrivenInstruction(scenario, messages, artifacts.getInteractionTrace());
+        String traceDrivenInstruction = instructionFactory.createTraceDrivenInstruction(scenario, artifacts.getInteractionTrace());
+        if (!traceDrivenInstruction.isEmpty()) {
+            messages.add(LLMChatMessage.user(traceDrivenInstruction));
+        }
         return Optional.empty();
     }
     
@@ -313,127 +266,6 @@ public final class LLMMCPConversationRunner {
         return Optional.empty();
     }
     
-    private void addTraceDrivenInstruction(final LLME2EScenario scenario, final List<LLMChatMessage> messages, final List<MCPInteractionTraceRecord> interactionTrace) {
-        if (interactionTrace.isEmpty()) {
-            return;
-        }
-        String resourceReadInstruction = createResourceReadInstruction(scenario, interactionTrace);
-        if (!resourceReadInstruction.isEmpty()) {
-            messages.add(LLMChatMessage.user(resourceReadInstruction));
-            return;
-        }
-        String immediateNextActionInstruction = createImmediateNextActionInstruction(interactionTrace.getLast());
-        if (!immediateNextActionInstruction.isEmpty()) {
-            messages.add(LLMChatMessage.user(immediateNextActionInstruction));
-            return;
-        }
-        if (hasSideEffectExecutionNextAction(interactionTrace)) {
-            messages.add(LLMChatMessage.user(createSideEffectExecutionNextActionInstruction(scenario.getExpectedAnswer())));
-        }
-    }
-    
-    private String createImmediateNextActionInstruction(final MCPInteractionTraceRecord traceRecord) {
-        for (Map<?, ?> each : LLMMCPNextActions.getNextActions(traceRecord.getStructuredContent())) {
-            String result = createMachineNextActionInstruction(each);
-            if (!result.isEmpty()) {
-                return result;
-            }
-        }
-        return "";
-    }
-    
-    private String findImmediateNextActionName(final MCPInteractionTraceRecord traceRecord) {
-        for (Map<?, ?> each : LLMMCPNextActions.getNextActions(traceRecord.getStructuredContent())) {
-            String result = findMachineNextActionName(each);
-            if (!result.isEmpty()) {
-                return result;
-            }
-        }
-        return "";
-    }
-    
-    private String findMachineNextActionName(final Map<?, ?> action) {
-        if (LLMMCPSideEffectNextAction.isExecutionAction(action)) {
-            return "";
-        }
-        String actionType = Objects.toString(action.get("type"), "").trim();
-        if ("resource_read".equals(actionType) && !Objects.toString(action.get("resource_uri"), "").trim().isEmpty()) {
-            return MCPInteractionActionNames.READ_RESOURCE;
-        }
-        if ("tool_call".equals(actionType)) {
-            return Objects.toString(action.get("tool_name"), "").trim();
-        }
-        return "completion".equals(actionType) ? MCPInteractionActionNames.COMPLETE : "";
-    }
-    
-    private String createMachineNextActionInstruction(final Map<?, ?> action) {
-        if (LLMMCPSideEffectNextAction.isExecutionAction(action)) {
-            return "";
-        }
-        String actionType = Objects.toString(action.get("type"), "").trim();
-        if ("resource_read".equals(actionType)) {
-            String resourceUri = Objects.toString(action.get("resource_uri"), "").trim();
-            return resourceUri.isEmpty()
-                    ? ""
-                    : String.format(Locale.ENGLISH,
-                            "The latest MCP response gave a read-only next_action. Call mcp_read_resource with uri `%s` now before any other MCP action or final answer.", resourceUri);
-        }
-        if ("tool_call".equals(actionType)) {
-            String toolName = Objects.toString(action.get("tool_name"), "").trim();
-            return toolName.isEmpty()
-                    ? ""
-                    : String.format(Locale.ENGLISH, "The latest MCP response gave an immediate next_action. Call `%s` now with exactly these arguments: %s. Do not replace values with placeholders.",
-                            toolName, JsonUtils.toJsonString(action.containsKey("arguments") ? action.get("arguments") : Map.of()));
-        }
-        if ("completion".equals(actionType)) {
-            Object arguments = action.containsKey("arguments") ? action.get("arguments") : createCompletionArguments(action);
-            return String.format(Locale.ENGLISH, "The latest MCP response gave an immediate completion next_action. Call mcp_complete now with exactly these arguments: %s.",
-                    JsonUtils.toJsonString(arguments));
-        }
-        return "";
-    }
-    
-    private Map<String, Object> createCompletionArguments(final Map<?, ?> action) {
-        Map<String, Object> result = new LinkedHashMap<>(3, 1F);
-        result.put("ref", action.get("ref"));
-        result.put("argument", action.get("argument"));
-        if (action.containsKey("context")) {
-            result.put("context", action.get("context"));
-        }
-        return result;
-    }
-    
-    private String createResourceReadInstruction(final LLME2EScenario scenario, final List<MCPInteractionTraceRecord> interactionTrace) {
-        if (!scenario.getRequiredToolNames().contains(MCPInteractionActionNames.READ_RESOURCE)
-                || !shouldPromptExactResourceRead(interactionTrace.getLast())) {
-            return "";
-        }
-        String resourceUri = LLMMCPScenarioInference.findExpectedResourceUri(scenario);
-        return resourceUri.isEmpty() || hasReadResource(resourceUri, interactionTrace)
-                ? ""
-                : String.format(Locale.ENGLISH,
-                        "The remaining required resource action is mcp_read_resource. Use exactly `%s` as uri; do not copy parameter schema or placeholder text as uri.", resourceUri);
-    }
-    
-    private boolean shouldPromptExactResourceRead(final MCPInteractionTraceRecord traceRecord) {
-        return MCPInteractionActionNames.LIST_RESOURCES.equals(traceRecord.getTargetName())
-                || MCPInteractionActionNames.READ_RESOURCE.equals(traceRecord.getTargetName())
-                        && (traceRecord.getStructuredContent().containsKey("error_code") || Boolean.FALSE.equals(traceRecord.getStructuredContent().get("found")));
-    }
-    
-    private boolean hasReadResource(final String resourceUri, final List<MCPInteractionTraceRecord> interactionTrace) {
-        for (MCPInteractionTraceRecord each : interactionTrace) {
-            if (MCPInteractionActionNames.RESOURCE_READ_KIND.equals(each.getActionKind()) && resourceUri.equals(each.getArguments().get("uri"))) {
-                return true;
-            }
-        }
-        return false;
-    }
-    
-    private boolean hasPendingImmediateNextAction(final List<MCPInteractionTraceRecord> interactionTrace) {
-        return !interactionTrace.isEmpty() && !createImmediateNextActionInstruction(interactionTrace.getLast()).isEmpty();
-    }
-    
     private Optional<LLME2EArtifactBundle> validateToolCall(final LLME2EScenario scenario, final String toolName, final Map<String, Object> arguments,
                                                             final LLMMCPConversationArtifacts artifacts) {
         return safetyValidator.validate(toolName, arguments)
@@ -473,116 +305,6 @@ public final class LLMMCPConversationRunner {
             Thread.currentThread().interrupt();
         } catch (final IOException ignored) {
         }
-    }
-    
-    private List<LLMChatMessage> createFinalAnswerMessages(final LLME2EScenario scenario, final List<MCPInteractionTraceRecord> interactionTrace) {
-        List<LLMChatMessage> result = new LinkedList<>();
-        result.add(LLMChatMessage.system("Return the final MCP assessment answer as valid JSON only."));
-        result.add(LLMChatMessage.user(createFinalAnswerInstruction(scenario, interactionTrace)));
-        return result;
-    }
-    
-    private String createFinalAnswerInstruction(final LLME2EScenario scenario, final List<MCPInteractionTraceRecord> interactionTrace) {
-        LLMStructuredAnswer expectedAnswer = scenario.getExpectedAnswer();
-        String interactionSequence = JsonUtils.toJsonString(createComparableInteractionSequence(interactionTrace));
-        String totalOrders = findLatestTotalOrders(interactionTrace);
-        String prompt = "Return JSON only with keys database, schema, table, query, totalOrders, interactionSequence. "
-                + "Use database `%s`, schema `%s`, table `%s`, and query `%s`; set totalOrders to `%s` from the latest successful database_gateway_execute_query result. "
-                + "Set interactionSequence exactly to this JSON array: %s. "
-                + "Do not add inferred, expected, available, or failed MCP action names. Required tools are `%s`.";
-        return String.format(Locale.ENGLISH,
-                prompt,
-                expectedAnswer.getDatabase(), expectedAnswer.getSchema(), expectedAnswer.getTable(), expectedAnswer.getQuery(), totalOrders, interactionSequence,
-                String.join(", ", scenario.getRequiredToolNames()));
-    }
-    
-    private String findLatestTotalOrders(final List<MCPInteractionTraceRecord> interactionTrace) {
-        for (int index = interactionTrace.size() - 1; index >= 0; index--) {
-            MCPInteractionTraceRecord each = interactionTrace.get(index);
-            if ("database_gateway_execute_query".equals(each.getTargetName()) && each.isValid()) {
-                String result = findTotalOrdersInRowObjects(each.getStructuredContent());
-                return result.isEmpty() ? findTotalOrdersInRows(each.getStructuredContent()) : result;
-            }
-        }
-        return "";
-    }
-    
-    private String findTotalOrdersInRowObjects(final Map<String, Object> structuredContent) {
-        List<Map<String, Object>> rowObjects = LLMMCPJsonValues.castToList(structuredContent.get("row_objects"));
-        if (rowObjects.isEmpty()) {
-            return "";
-        }
-        return Objects.toString(rowObjects.getFirst().get("total_orders"), "").trim();
-    }
-    
-    private String findTotalOrdersInRows(final Map<String, Object> structuredContent) {
-        List<Object> rows = LLMMCPJsonValues.castToList(structuredContent.get("rows"));
-        if (rows.isEmpty()) {
-            return "";
-        }
-        List<Object> row = LLMMCPJsonValues.castToList(rows.getFirst());
-        return row.isEmpty() ? "" : Objects.toString(row.getFirst(), "").trim();
-    }
-    
-    private List<String> createComparableInteractionSequence(final List<MCPInteractionTraceRecord> interactionTrace) {
-        List<String> result = new LinkedList<>();
-        for (MCPInteractionTraceRecord each : interactionTrace) {
-            if (result.isEmpty() || !result.getLast().equals(each.getTargetName())) {
-                result.add(each.getTargetName());
-            }
-        }
-        return result;
-    }
-    
-    private String createRequiredToolCallInstruction(final LLME2EScenario scenario, final LLMMCPConversationArtifacts artifacts) {
-        List<String> missingToolNames = LLMMCPInteractionCoverage.findMissingRequiredInteractionNames(scenario.getRequiredToolNames(), artifacts.getInteractionTrace());
-        LLMStructuredAnswer expectedAnswer = scenario.getExpectedAnswer();
-        String previewInstruction = missingToolNames.contains("database_gateway_execute_update")
-                ? String.format(Locale.ENGLISH,
-                        " For database_gateway_execute_update, set database `%s`, schema `%s`, execution_mode=preview, and keep the side-effecting SQL unchanged; do not use execution_mode=execute.",
-                        expectedAnswer.getDatabase(), expectedAnswer.getSchema())
-                : "";
-        String resourceInstruction = missingToolNames.contains(MCPInteractionActionNames.READ_RESOURCE)
-                ? " For mcp_read_resource, use an exact shardingsphere:// URI from the user request or the latest tool response; do not invent abbreviated URI strings."
-                : "";
-        String planningInstruction = hasMissingPlanningTool(missingToolNames)
-                ? " For a new database_gateway_plan_* call, omit plan_id unless a previous MCP planning response returned an actual plan_id."
-                : "";
-        String workflowPlanInstruction = createWorkflowPlanInstruction(missingToolNames, artifacts.getInteractionTrace());
-        return String.format(Locale.ENGLISH,
-                "Required MCP tool coverage is incomplete. Remaining required MCP tools: %s. "
-                        + "Call one remaining tool as an actual MCP tool_call now; do not answer in text, do not write JSON, and do not write <tool_call> markup. "
-                        + "If database_gateway_execute_query is remaining, set database `%s`, schema `%s`, and sql `%s`.%s%s%s%s",
-                String.join(", ", missingToolNames), expectedAnswer.getDatabase(), expectedAnswer.getSchema(), expectedAnswer.getQuery(), previewInstruction, resourceInstruction,
-                planningInstruction, workflowPlanInstruction);
-    }
-    
-    private boolean hasMissingPlanningTool(final List<String> missingToolNames) {
-        return missingToolNames.stream().anyMatch(each -> each.startsWith(LLMMCPScenarioInference.PLANNING_TOOL_NAME_PREFIX));
-    }
-    
-    private String createSideEffectExecutionNextActionInstruction(final LLMStructuredAnswer expectedAnswer) {
-        return String.format(Locale.ENGLISH,
-                "The latest MCP response contains side-effect execution next_actions; do not execute them in this score lane. "
-                        + "Call database_gateway_execute_query now with database `%s`, schema `%s`, and sql `%s`. "
-                        + "Do not call database_gateway_execute_update for SELECT or row-count verification.",
-                expectedAnswer.getDatabase(), expectedAnswer.getSchema(), expectedAnswer.getQuery());
-    }
-    
-    private boolean hasSideEffectExecutionNextAction(final List<MCPInteractionTraceRecord> interactionTrace) {
-        return !interactionTrace.isEmpty() && LLMMCPNextActions.getNextActions(interactionTrace.getLast().getStructuredContent()).stream().anyMatch(LLMMCPSideEffectNextAction::isExecutionAction);
-    }
-    
-    private String createWorkflowPlanInstruction(final List<String> missingToolNames, final List<MCPInteractionTraceRecord> interactionTrace) {
-        if (!missingToolNames.contains("database_gateway_apply_workflow") && !missingToolNames.contains("database_gateway_validate_workflow")) {
-            return "";
-        }
-        String latestPlanId = LLMMCPScenarioInference.findLatestPlanId(interactionTrace);
-        return latestPlanId.isEmpty()
-                ? " For database_gateway_apply_workflow or database_gateway_validate_workflow, use an actual plan_id returned by a successful planning tool call; "
-                        + "do not use placeholder text `plan_id`."
-                : String.format(Locale.ENGLISH,
-                        " For database_gateway_apply_workflow or database_gateway_validate_workflow, set plan_id `%s`; do not use placeholder text `plan_id`.", latestPlanId);
     }
     
     private MCPInteractionTraceRecord createTraceRecord(final int sequence, final String actionName, final String actionOrigin, final Map<String, Object> arguments,
